Happy Eyeballs and fallback for the proxy hostname
When using a proxy that is specified by hostname, this change causes the
client to try all IPs associated with that hostname until it finds one
that works.  (It will also prefer IPv6 addresses if they are present.)
If an IP stops working, the client will try the other ones, and repeat
the name resolution if necessary.

This change also copies IPSet from outline-go-tun2socks/doh/intra/ipmap
(with some additions) which will subsequently be removed there.

Related issues:
https://github.com/Jigsaw-Code/outline-client/issues/91
https://github.com/Jigsaw-Code/outline-client/issues/671
https://github.com/Jigsaw-Code/outline-server/issues/501