Use the IP PacketConn to specify the local proxy IP
This allows the proxy to determine the destination address of incoming
UDP packets and specify the source address of outgoing UDP packets.  It
requires duplicating the UDP service because it uses the x/net/ipv[4,6]
packages, which require us to know the address family of each incoming
packet before it is received.